Fixed crash when reordering
